removed F5diagnostics listener
